moved outputs to appropriate package
